requirements.txt
bleach
fpdf2
pikepdf
psycopg[binary]
pyyaml